moved write method into threemf_package
